Moved deprecated components to "deprecated" dir
