+++ packages/asc-ui/src/components/deprecated/ButtonBar/ButtonBar.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { shallow } from 'enzyme'
-import Button from '../Button'
+import Button from '../../Button'
 import ButtonBar from './ButtonBar'
 import 'jest-styled-components'
