+++ packages/asc-ui/src/index.ts
@@ -2,7 +2,6 @@ import { ServerStyleSheet, Theme, ascDefaultTheme } from '@datapunt/asc-core'
 
 import AmsterdamLogo, { AmsterdamLogoStyles } from './components/AmsterdamLogo'
 import BackDrop, { BackDropStyles } from './components/BackDrop'
-import IconButton from './components/Button/IconButton'
 import Icon from './components/Icon'
 import Input, { InputStyles } from './components/Input'
 import Modal from './components/Modal'
@@ -67,7 +66,8 @@ import Spinner from './components/Spinner/Spinner'
 import LinkList, { LinkListItem } from './components/LinkList'
 import DocumentCover from './components/DocumentCover'
 import { Article, ArticleStyle } from './components/Editorial/Article'
-import ButtonBar from './components/ButtonBar'
+import IconButton from './components/deprecated/IconButton'
+import ButtonBar from './components/deprecated/ButtonBar'
 
 export { breakpoint } from './utils'
